feat: add TicketSoldOutBottomSheet
